# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GlobeTrotter API
  version: 0.1.0
  description: |
    Phase 1 API contract for GlobeTrotter.
    Downstream phases should build against these shapes before live handlers exist.

servers:
  - url: http://localhost:8000/api/v1
    description: Local development

tags:
  - name: auth
  - name: users
  - name: trips
  - name: cities
  - name: activities
  - name: stops
  - name: trip-activities
  - name: shares

paths:
  /auth/signup:
    post:
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/forgot-password:
    post:
      tags: [auth]
      summary: Request password reset (stub)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email queued (stub)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  /auth/me:
    get:
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /users/me:
    get:
      tags: [users]
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [users]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRead'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /trips:
    get:
      tags: [trips]
      summary: List trips for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripRead'
    post:
      tags: [trips]
      summary: Create a trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Trip created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripRead'

  /trips/{trip_id}:
    parameters:
      - $ref: '#/components/parameters/TripId'
    get:
      tags: [trips]
      summary: Get trip by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripRead'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [trips]
      summary: Update trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripUpdate'
      responses:
        '200':
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripRead'
    delete:
      tags: [trips]
      summary: Delete trip
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /cities:
    get:
      tags: [cities]
      summary: List or search cities
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
      responses:
        '200':
          description: City list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CityRead'
    post:
      tags: [cities]
      summary: Create city (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityCreate'
      responses:
        '201':
          description: City created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityRead'

  /cities/{city_id}:
    parameters:
      - $ref: '#/components/parameters/CityId'
    get:
      tags: [cities]
      summary: Get city by ID
      responses:
        '200':
          description: City
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityRead'
    patch:
      tags: [cities]
      summary: Update city (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CityUpdate'
      responses:
        '200':
          description: Updated city
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityRead'
    delete:
      tags: [cities]
      summary: Delete city (admin)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /activities:
    get:
      tags: [activities]
      summary: List or filter activities
      parameters:
        - name: city_id
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityRead'
    post:
      tags: [activities]
      summary: Create activity (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityCreate'
      responses:
        '201':
          description: Activity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityRead'

  /activities/{activity_id}:
    parameters:
      - $ref: '#/components/parameters/ActivityId'
    get:
      tags: [activities]
      summary: Get activity by ID
      responses:
        '200':
          description: Activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityRead'
    patch:
      tags: [activities]
      summary: Update activity (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityUpdate'
      responses:
        '200':
          description: Updated activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityRead'
    delete:
      tags: [activities]
      summary: Delete activity (admin)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /stops:
    get:
      tags: [stops]
      summary: List stops
      parameters:
        - name: trip_id
          in: query
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stop list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StopRead'
    post:
      tags: [stops]
      summary: Create stop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopCreate'
      responses:
        '201':
          description: Stop created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopRead'

  /stops/{stop_id}:
    parameters:
      - $ref: '#/components/parameters/StopId'
    get:
      tags: [stops]
      summary: Get stop by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Stop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopRead'
    patch:
      tags: [stops]
      summary: Update stop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopUpdate'
      responses:
        '200':
          description: Updated stop
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StopRead'
    delete:
      tags: [stops]
      summary: Delete stop
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /trip-activities:
    get:
      tags: [trip-activities]
      summary: List trip activities
      parameters:
        - name: stop_id
          in: query
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Trip activity list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripActivityRead'
    post:
      tags: [trip-activities]
      summary: Schedule activity on a stop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripActivityCreate'
      responses:
        '201':
          description: Trip activity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripActivityRead'

  /trip-activities/{trip_activity_id}:
    parameters:
      - $ref: '#/components/parameters/TripActivityId'
    get:
      tags: [trip-activities]
      summary: Get trip activity by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Trip activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripActivityRead'
    patch:
      tags: [trip-activities]
      summary: Update trip activity
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripActivityUpdate'
      responses:
        '200':
          description: Updated trip activity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripActivityRead'
    delete:
      tags: [trip-activities]
      summary: Delete trip activity
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

  /shares:
    post:
      tags: [shares]
      summary: Create public share link for a trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripShareCreate'
      responses:
        '201':
          description: Share created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripShareRead'

  /shares/public/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [shares]
      summary: Get public share metadata by slug
      responses:
        '200':
          description: Share metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripShareRead'
        '404':
          $ref: '#/components/responses/NotFound'

  /shares/{share_id}:
    parameters:
      - name: share_id
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags: [shares]
      summary: Revoke share link
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    TripId:
      name: trip_id
      in: path
      required: true
      schema:
        type: integer
    CityId:
      name: city_id
      in: path
      required: true
      schema:
        type: integer
    ActivityId:
      name: activity_id
      in: path
      required: true
      schema:
        type: integer
    StopId:
      name: stop_id
      in: path
      required: true
      schema:
        type: integer
    TripActivityId:
      name: trip_activity_id
      in: path
      required: true
      schema:
        type: integer

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - type: array
              items:
                type: object

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    SignupRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    TokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer

    UserRead:
      type: object
      required: [id, name, email, language_pref, is_admin, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        photo:
          type: string
          nullable: true
        language_pref:
          type: string
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        photo:
          type: string
          nullable: true
        language_pref:
          type: string

    TripCreate:
      type: object
      required: [name, start_date, end_date]
      properties:
        name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        cover_photo:
          type: string
          nullable: true
        is_public:
          type: boolean
          default: false
        budget_cap:
          type: string
          format: decimal
          nullable: true

    TripUpdate:
      type: object
      properties:
        name:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        description:
          type: string
          nullable: true
        cover_photo:
          type: string
          nullable: true
        is_public:
          type: boolean
        budget_cap:
          type: string
          format: decimal
          nullable: true

    TripRead:
      allOf:
        - $ref: '#/components/schemas/TripCreate'
        - type: object
          required: [id, user_id, created_at]
          properties:
            id:
              type: integer
            user_id:
              type: integer
            created_at:
              type: string
              format: date-time

    CityCreate:
      type: object
      required: [name, country, cost_index]
      properties:
        name:
          type: string
        country:
          type: string
        cost_index:
          type: string
          format: decimal
        popularity_score:
          type: integer
          default: 0
        image_url:
          type: string
          nullable: true

    CityUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        cost_index:
          type: string
          format: decimal
        popularity_score:
          type: integer
        image_url:
          type: string
          nullable: true

    CityRead:
      allOf:
        - $ref: '#/components/schemas/CityCreate'
        - type: object
          required: [id]
          properties:
            id:
              type: integer

    ActivityCreate:
      type: object
      required: [city_id, name, category, cost, duration]
      properties:
        city_id:
          type: integer
        name:
          type: string
        category:
          type: string
        cost:
          type: string
          format: decimal
        duration:
          type: integer
          minimum: 1
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true

    ActivityUpdate:
      type: object
      properties:
        city_id:
          type: integer
        name:
          type: string
        category:
          type: string
        cost:
          type: string
          format: decimal
        duration:
          type: integer
        description:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true

    ActivityRead:
      allOf:
        - $ref: '#/components/schemas/ActivityCreate'
        - type: object
          required: [id]
          properties:
            id:
              type: integer

    StopCreate:
      type: object
      required: [trip_id, city_id, start_date, end_date, order_index]
      properties:
        trip_id:
          type: integer
        city_id:
          type: integer
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        order_index:
          type: integer

    StopUpdate:
      type: object
      properties:
        city_id:
          type: integer
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        order_index:
          type: integer

    StopRead:
      allOf:
        - $ref: '#/components/schemas/StopCreate'
        - type: object
          required: [id]
          properties:
            id:
              type: integer

    TripActivityCreate:
      type: object
      required: [stop_id, activity_id, scheduled_date]
      properties:
        stop_id:
          type: integer
        activity_id:
          type: integer
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
          format: time
          nullable: true
        cost_override:
          type: string
          format: decimal
          nullable: true

    TripActivityUpdate:
      type: object
      properties:
        activity_id:
          type: integer
        scheduled_date:
          type: string
          format: date
        scheduled_time:
          type: string
          format: time
          nullable: true
        cost_override:
          type: string
          format: decimal
          nullable: true

    TripActivityRead:
      allOf:
        - $ref: '#/components/schemas/TripActivityCreate'
        - type: object
          required: [id]
          properties:
            id:
              type: integer

    TripShareCreate:
      type: object
      required: [trip_id]
      properties:
        trip_id:
          type: integer

    TripShareRead:
      type: object
      required: [id, trip_id, public_slug, created_at]
      properties:
        id:
          type: integer
        trip_id:
          type: integer
        public_slug:
          type: string
        created_at:
          type: string
          format: date-time